+++ opentelemetry-sdk/src/trace/sampler.rs
@@ -2,7 +2,7 @@ use opentelemetry::{
     trace::{
         Link, SamplingDecision, SamplingResult, SpanKind, TraceContextExt, TraceId, TraceState,
     },
-    Context, Key, OrderMap, Value,
+    Context, KeyValue,
 };
 use std::convert::TryInto;
 
@@ -78,7 +78,7 @@ pub trait ShouldSample: CloneShouldSample + Send + Sync + std::fmt::Debug {
         trace_id: TraceId,
         name: &str,
         span_kind: &SpanKind,
-        attributes: &OrderMap<Key, Value>,
+        attributes: &[KeyValue],
         links: &[Link],
     ) -> SamplingResult;
 }
@@ -170,7 +170,7 @@ impl ShouldSample for Sampler {
         trace_id: TraceId,
         name: &str,
         span_kind: &SpanKind,
-        attributes: &OrderMap<Key, Value>,
+        attributes: &[KeyValue],
         links: &[Link],
     ) -> SamplingResult {
         let decision = match self {
@@ -333,7 +333,7 @@ mod tests {
                         trace_id,
                         name,
                         &SpanKind::Internal,
-                        &Default::default(),
+                        &[],
                         &[],
                     )
                     .decision
@@ -377,7 +377,7 @@ mod tests {
             TraceId::from_u128(1),
             "should sample",
             &SpanKind::Internal,
-            &Default::default(),
+            &[],
             &[],
         );
 
@@ -386,7 +386,7 @@ mod tests {
             TraceId::from_u128(1),
             "should sample",
             &SpanKind::Internal,
-            &Default::default(),
+            &[],
             &[],
         );
 
@@ -436,7 +436,7 @@ mod tests {
                 TraceId::from_u128(1),
                 name,
                 &SpanKind::Internal,
-                &Default::default(),
+                &[],
                 &[],
             );