+++ opentelemetry/src/trace/tracer.rs
@@ -1,9 +1,8 @@
 use crate::{
     trace::{Event, Link, Span, SpanId, SpanKind, Status, TraceContextExt, TraceId, TraceState},
-    Context, Key, KeyValue, OrderMap, Value,
+    Context, KeyValue,
 };
 use std::borrow::Cow;
-use std::iter::FromIterator;
 use std::time::SystemTime;
 
 /// The interface for constructing [`Span`]s.
@@ -259,8 +258,11 @@ pub struct SpanBuilder {
     /// Span end time
     pub end_time: Option<SystemTime>,
 
-    /// Span attributes
-    pub attributes: Option<OrderMap<Key, Value>>,
+    /// Span attributes that are provided at the span creation time.
+    /// More attributes can be added afterwards.
+    /// Providing duplicate keys will result in multiple attributes
+    /// with the same key, as there is no de-duplication performed.
+    pub attributes: Option<Vec<KeyValue>>,
 
     /// Span events
     pub events: Option<Vec<Event>>,
@@ -326,26 +328,14 @@ impl SpanBuilder {
     }
 
     /// Assign span attributes from an iterable.
-    ///
-    /// Check out [`SpanBuilder::with_attributes_map`] to assign span attributes
-    /// via an [`OrderMap`] instance.
+    /// Providing duplicate keys will result in multiple attributes
+    /// with the same key, as there is no de-duplication performed.    
     pub fn with_attributes<I>(self, attributes: I) -> Self
     where
         I: IntoIterator<Item = KeyValue>,
     {
         SpanBuilder {
-            attributes: Some(OrderMap::from_iter(attributes)),
-            ..self
-        }
-    }
-
-    /// Assign span attributes.
-    ///
-    /// Check out [`SpanBuilder::with_attributes`] to assign span attributes
-    /// from an iterable of [`KeyValue`]s.
-    pub fn with_attributes_map(self, attributes: OrderMap<Key, Value>) -> Self {
-        SpanBuilder {
-            attributes: Some(attributes),
+            attributes: Some(attributes.into_iter().collect()),
             ..self
         }
     }